refactor: 방생성화면 코드 영역 별도 패키지로 분리
